+++ tests/encoding.c
@@ -7,25 +7,25 @@
 #include "test.h"
 
 void *test_encoding(void *arg) {
-    ATT_ASSERT(mjb_string_encoding(0, 10), MJB_ENCODING_UNKNOWN, "Void unknown string")
-    ATT_ASSERT(mjb_string_encoding("", 0), MJB_ENCODING_UNKNOWN, "Void unknown length");
+    ATT_ASSERT((unsigned int)mjb_string_encoding(0, 10), (unsigned int)MJB_ENCODING_UNKNOWN, "Void unknown string")
+    ATT_ASSERT((unsigned int)mjb_string_encoding("", 0), (unsigned int)MJB_ENCODING_UNKNOWN, "Void unknown length");
 
-    ATT_ASSERT(mjb_string_encoding(0, 0), MJB_ENCODING_UNKNOWN, "Void unknown string and length");
+    ATT_ASSERT((unsigned int)mjb_string_encoding(0, 0), (unsigned int)MJB_ENCODING_UNKNOWN, "Void unknown string and length");
     const char *test1 = "The quick brown fox jumps over the lazy dog";
 
     ATT_ASSERT(mjb_string_encoding(test1, 43), (MJB_ENCODING_ASCII | MJB_ENCODING_UTF_8), "Plain ASCII (and UTF-8)");
     const char *test2 = "\xEF\xBB\xBFThe quick brown fox jumps over the lazy dog";
 
-    ATT_ASSERT(mjb_string_encoding(test2, 43 + 3), MJB_ENCODING_UTF_8, "UTF-8 BOM");
+    ATT_ASSERT((unsigned int)mjb_string_encoding(test2, 43 + 3), (unsigned int)MJB_ENCODING_UTF_8, "UTF-8 BOM");
     const char *test3 = "\xFE\xFFThe quick brown fox jumps over the lazy dog";
 
-    ATT_ASSERT(mjb_string_encoding(test3, 43 + 2), MJB_ENCODING_UTF_16 | MJB_ENCODING_UTF_16_BE, "UTF-16-BE BOM");
+    ATT_ASSERT((unsigned int)mjb_string_encoding(test3, 43 + 2), (unsigned int)MJB_ENCODING_UTF_16 | MJB_ENCODING_UTF_16_BE, "UTF-16-BE BOM");
 
     const char *test4 = "\xFF\xFEThe quick brown fox jumps over the lazy dog";
-    ATT_ASSERT(mjb_string_encoding(test4, 43 + 2), MJB_ENCODING_UTF_16 | MJB_ENCODING_UTF_16_LE, "UTF-16-LE BOM");
+    ATT_ASSERT((unsigned int)mjb_string_encoding(test4, 43 + 2), (unsigned int)MJB_ENCODING_UTF_16 | MJB_ENCODING_UTF_16_LE, "UTF-16-LE BOM");
 
     const char *test5 = "\x00\x00\xFE\xFFThe quick brown fox jumps over the lazy dog";
-    ATT_ASSERT(mjb_string_encoding(test5, 43 + 4), MJB_ENCODING_UTF_32 | MJB_ENCODING_UTF_32_BE, "UTF-32-BE BOM");
+    ATT_ASSERT((unsigned int)mjb_string_encoding(test5, 43 + 4), (unsigned int)MJB_ENCODING_UTF_32 | MJB_ENCODING_UTF_32_BE, "UTF-32-BE BOM");
 
     const char *test6 = "\xFF\xFE\x00\x00The quick brown fox jumps over the lazy dog";
     ATT_ASSERT(mjb_string_encoding(test6, 43 + 4), MJB_ENCODING_UTF_32 | MJB_ENCODING_UTF_32_LE | MJB_ENCODING_UTF_16 | MJB_ENCODING_UTF_16_LE, "UTF-32-LE BOM");
